mlst now handles where true klebsiella has equal scheme/mlst correctness scores to ecoli(achtman)

+++ modules/local/mlst.nf
@@ -71,10 +71,15 @@ process MLST {
         cat ${prefix}_1.tsv ${prefix}_2.tsv > ${prefix}.tsv
         rm ${prefix}_*.tsv
     elif [[ \$scheme == "ecoli_achtman_4" ]]; then
-        mv ${prefix}.tsv ${prefix}_1.tsv
-        mlst --scheme ecoli --threads $task.cpus \$unzipped_fasta > ${prefix}_2.tsv || mlst --scheme ecoli_2 --threads $task.cpus \$unzipped_fasta > ${prefix}_2.tsv
-        cat ${prefix}_1.tsv ${prefix}_2.tsv > ${prefix}.tsv
-        rm ${prefix}_*.tsv
+        if [[ ${genus,,} == "klebsiella" ]]; then
+            mv ${prefix}.tsv ${prefix}.OLD-tsv
+            mlst --scheme klebsiella --threads 4 $unzipped_fasta > ${prefix}.tsv
+        else:
+            mv ${prefix}.tsv ${prefix}_1.tsv
+            mlst --scheme ecoli --threads $task.cpus \$unzipped_fasta > ${prefix}_2.tsv || mlst --scheme ecoli_2 --threads $task.cpus \$unzipped_fasta > ${prefix}_2.tsv
+            cat ${prefix}_1.tsv ${prefix}_2.tsv > ${prefix}.tsv
+            rm ${prefix}_*.tsv
+        fi
     elif [[ \${genus,,} == "mycobacterium" ]]; then
         if [[ \$scheme == "mabscessus" ]]; then
             mv ${prefix}.tsv ${prefix}_1.tsv
